moved tests out of ibc hooks
