+++ tests/ibc-hooks/testutils/contracts/counter/src/contract.rs
