+++ tests/ibc-hooks/testutils/contracts/counter/src/helpers.rs
