--- x/ibc-hooks/ibc_middleware_test.go
@@ -1,464 +1,0 @@
-package ibc_hooks_test
-
-import (
-	"encoding/json"
-	"fmt"
-	"testing"
-
-	ibchooks "github.com/osmosis-labs/osmosis/v13/x/ibc-hooks"
-
-	"github.com/osmosis-labs/osmosis/v13/osmoutils"
-
-	"github.com/osmosis-labs/osmosis/v13/app/apptesting"
-
-	"github.com/stretchr/testify/suite"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	transfertypes "github.com/cosmos/ibc-go/v3/modules/apps/transfer/types"
-	clienttypes "github.com/cosmos/ibc-go/v3/modules/core/02-client/types"
-	channeltypes "github.com/cosmos/ibc-go/v3/modules/core/04-channel/types"
-	ibctesting "github.com/cosmos/ibc-go/v3/testing"
-
-	osmosisibctesting "github.com/osmosis-labs/osmosis/v13/x/ibc-rate-limit/testutil"
-
-	"github.com/osmosis-labs/osmosis/v13/x/ibc-hooks/testutils"
-)
-
-type HooksTestSuite struct {
-	apptesting.KeeperTestHelper
-
-	coordinator *ibctesting.Coordinator
-
-	chainA *osmosisibctesting.TestChain
-	chainB *osmosisibctesting.TestChain
-
-	path *ibctesting.Path
-}
-
-func (suite *HooksTestSuite) SetupTest() {
-	suite.Setup()
-	ibctesting.DefaultTestingAppInit = osmosisibctesting.SetupTestingApp
-	suite.coordinator = ibctesting.NewCoordinator(suite.T(), 2)
-	suite.chainA = &osmosisibctesting.TestChain{
-		TestChain: suite.coordinator.GetChain(ibctesting.GetChainID(1)),
-	}
-	suite.chainB = &osmosisibctesting.TestChain{
-		TestChain: suite.coordinator.GetChain(ibctesting.GetChainID(2)),
-	}
-	err := suite.chainA.MoveEpochsToTheFuture()
-	suite.Require().NoError(err)
-	err = suite.chainB.MoveEpochsToTheFuture()
-	suite.Require().NoError(err)
-	suite.path = NewTransferPath(suite.chainA, suite.chainB)
-	suite.coordinator.Setup(suite.path)
-}
-
-func TestIBCHooksTestSuite(t *testing.T) {
-	suite.Run(t, new(HooksTestSuite))
-}
-
-// ToDo: Move this to osmosistesting to avoid repetition
-func NewTransferPath(chainA, chainB *osmosisibctesting.TestChain) *ibctesting.Path {
-	path := ibctesting.NewPath(chainA.TestChain, chainB.TestChain)
-	path.EndpointA.ChannelConfig.PortID = ibctesting.TransferPort
-	path.EndpointB.ChannelConfig.PortID = ibctesting.TransferPort
-	path.EndpointA.ChannelConfig.Version = transfertypes.Version
-	path.EndpointB.ChannelConfig.Version = transfertypes.Version
-
-	return path
-}
-
-func (suite *HooksTestSuite) TestOnRecvPacketHooks() {
-	var (
-		trace    transfertypes.DenomTrace
-		amount   sdk.Int
-		receiver string
-		status   testutils.Status
-	)
-
-	testCases := []struct {
-		msg      string
-		malleate func(*testutils.Status)
-		expPass  bool
-	}{
-		{"override", func(status *testutils.Status) {
-			suite.chainB.GetOsmosisApp().TransferStack.
-				ICS4Middleware.Hooks = testutils.TestRecvOverrideHooks{Status: status}
-		}, true},
-		{"before and after", func(status *testutils.Status) {
-			suite.chainB.GetOsmosisApp().TransferStack.
-				ICS4Middleware.Hooks = testutils.TestRecvBeforeAfterHooks{Status: status}
-		}, true},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-		suite.Run(tc.msg, func() {
-			suite.SetupTest() // reset
-
-			path := NewTransferPath(suite.chainA, suite.chainB)
-			suite.coordinator.Setup(path)
-			receiver = suite.chainB.SenderAccount.GetAddress().String() // must be explicitly changed in malleate
-			status = testutils.Status{}
-
-			amount = sdk.NewInt(100) // must be explicitly changed in malleate
-			seq := uint64(1)
-
-			trace = transfertypes.ParseDenomTrace(sdk.DefaultBondDenom)
-
-			// send coin from chainA to chainB
-			transferMsg := transfertypes.NewMsgTransfer(path.EndpointA.ChannelConfig.PortID, path.EndpointA.ChannelID, sdk.NewCoin(trace.IBCDenom(), amount), suite.chainA.SenderAccount.GetAddress().String(), receiver, clienttypes.NewHeight(1, 110), 0)
-			_, err := suite.chainA.SendMsgs(transferMsg)
-			suite.Require().NoError(err) // message committed
-
-			tc.malleate(&status)
-
-			data := transfertypes.NewFungibleTokenPacketData(trace.GetFullDenomPath(), amount.String(), suite.chainA.SenderAccount.GetAddress().String(), receiver)
-			packet := channeltypes.NewPacket(data.GetBytes(), seq, path.EndpointA.ChannelConfig.PortID, path.EndpointA.ChannelID, path.EndpointB.ChannelConfig.PortID, path.EndpointB.ChannelID, clienttypes.NewHeight(1, 100), 0)
-
-			ack := suite.chainB.GetOsmosisApp().TransferStack.
-				OnRecvPacket(suite.chainB.GetContext(), packet, suite.chainA.SenderAccount.GetAddress())
-
-			if tc.expPass {
-				suite.Require().True(ack.Success())
-			} else {
-				suite.Require().False(ack.Success())
-			}
-
-			if _, ok := suite.chainB.GetOsmosisApp().TransferStack.
-				ICS4Middleware.Hooks.(testutils.TestRecvOverrideHooks); ok {
-				suite.Require().True(status.OverrideRan)
-				suite.Require().False(status.BeforeRan)
-				suite.Require().False(status.AfterRan)
-			}
-
-			if _, ok := suite.chainB.GetOsmosisApp().TransferStack.
-				ICS4Middleware.Hooks.(testutils.TestRecvBeforeAfterHooks); ok {
-				suite.Require().False(status.OverrideRan)
-				suite.Require().True(status.BeforeRan)
-				suite.Require().True(status.AfterRan)
-			}
-		})
-	}
-}
-
-func (suite *HooksTestSuite) makeMockPacket(receiver, memo string, prevSequence uint64) channeltypes.Packet {
-	packetData := transfertypes.FungibleTokenPacketData{
-		Denom:    sdk.DefaultBondDenom,
-		Amount:   "1",
-		Sender:   suite.chainB.SenderAccount.GetAddress().String(),
-		Receiver: receiver,
-		Memo:     memo,
-	}
-
-	return channeltypes.NewPacket(
-		packetData.GetBytes(),
-		prevSequence+1,
-		suite.path.EndpointB.ChannelConfig.PortID,
-		suite.path.EndpointB.ChannelID,
-		suite.path.EndpointA.ChannelConfig.PortID,
-		suite.path.EndpointA.ChannelID,
-		clienttypes.NewHeight(0, 100),
-		0,
-	)
-}
-
-func (suite *HooksTestSuite) receivePacket(receiver, memo string) []byte {
-	return suite.receivePacketWithSequence(receiver, memo, 0)
-}
-
-func (suite *HooksTestSuite) receivePacketWithSequence(receiver, memo string, prevSequence uint64) []byte {
-	channelCap := suite.chainB.GetChannelCapability(
-		suite.path.EndpointB.ChannelConfig.PortID,
-		suite.path.EndpointB.ChannelID)
-
-	packet := suite.makeMockPacket(receiver, memo, prevSequence)
-
-	err := suite.chainB.GetOsmosisApp().HooksICS4Wrapper.SendPacket(
-		suite.chainB.GetContext(), channelCap, packet)
-	suite.Require().NoError(err, "IBC send failed. Expected success. %s", err)
-
-	// Update both clients
-	err = suite.path.EndpointB.UpdateClient()
-	suite.Require().NoError(err)
-	err = suite.path.EndpointA.UpdateClient()
-	suite.Require().NoError(err)
-
-	// recv in chain a
-	res, err := suite.path.EndpointA.RecvPacketWithResult(packet)
-
-	// get the ack from the chain a's response
-	ack, err := ibctesting.ParseAckFromEvents(res.GetEvents())
-	suite.Require().NoError(err)
-
-	// manually send the acknowledgement to chain b
-	err = suite.path.EndpointA.AcknowledgePacket(packet, ack)
-	suite.Require().NoError(err)
-	return ack
-}
-
-func (suite *HooksTestSuite) TestRecvTransferWithMetadata() {
-	// Setup contract
-	suite.chainA.StoreContractCode(&suite.Suite, "./bytecode/echo.wasm")
-	addr := suite.chainA.InstantiateContract(&suite.Suite, "{}", 1)
-
-	ackBytes := suite.receivePacket(addr.String(), fmt.Sprintf(`{"wasm": {"contract": "%s", "msg": {"echo": {"msg": "test"} } } }`, addr))
-	ackStr := string(ackBytes)
-	fmt.Println(ackStr)
-	var ack map[string]string // This can't be unmarshalled to Acknowledgement because it's fetched from the events
-	err := json.Unmarshal(ackBytes, &ack)
-	suite.Require().NoError(err)
-	suite.Require().NotContains(ack, "error")
-	suite.Require().Equal(ack["result"], "eyJjb250cmFjdF9yZXN1bHQiOiJkR2hwY3lCemFHOTFiR1FnWldOb2J3PT0iLCJpYmNfYWNrIjoiZXlKeVpYTjFiSFFpT2lKQlVUMDlJbjA9In0=")
-}
-
-// After successfully executing a wasm call, the contract should have the funds sent via IBC
-func (suite *HooksTestSuite) TestFundsAreTransferredToTheContract() {
-	// Setup contract
-	suite.chainA.StoreContractCode(&suite.Suite, "./bytecode/echo.wasm")
-	addr := suite.chainA.InstantiateContract(&suite.Suite, "{}", 1)
-
-	// Check that the contract has no funds
-	localDenom := osmoutils.MustExtractDenomFromPacketOnRecv(suite.makeMockPacket("", "", 0))
-	balance := suite.chainA.GetOsmosisApp().BankKeeper.GetBalance(suite.chainA.GetContext(), addr, localDenom)
-	suite.Require().Equal(sdk.NewInt(0), balance.Amount)
-
-	// Execute the contract via IBC
-	ackBytes := suite.receivePacket(addr.String(), fmt.Sprintf(`{"wasm": {"contract": "%s", "msg": {"echo": {"msg": "test"} } } }`, addr))
-	ackStr := string(ackBytes)
-	fmt.Println(ackStr)
-	var ack map[string]string // This can't be unmarshalled to Acknowledgement because it's fetched from the events
-	err := json.Unmarshal(ackBytes, &ack)
-	suite.Require().NoError(err)
-	suite.Require().NotContains(ack, "error")
-	suite.Require().Equal(ack["result"], "eyJjb250cmFjdF9yZXN1bHQiOiJkR2hwY3lCemFHOTFiR1FnWldOb2J3PT0iLCJpYmNfYWNrIjoiZXlKeVpYTjFiSFFpT2lKQlVUMDlJbjA9In0=")
-
-	// Check that the token has now been transferred to the contract
-	balance = suite.chainA.GetOsmosisApp().BankKeeper.GetBalance(suite.chainA.GetContext(), addr, localDenom)
-	suite.Require().Equal(sdk.NewInt(1), balance.Amount)
-}
-
-// If the wasm call wails, the contract acknowledgement should be an error and the funds returned
-func (suite *HooksTestSuite) TestFundsAreReturnedOnFailedContractExec() {
-	// Setup contract
-	suite.chainA.StoreContractCode(&suite.Suite, "./bytecode/echo.wasm")
-	addr := suite.chainA.InstantiateContract(&suite.Suite, "{}", 1)
-
-	// Check that the contract has no funds
-	localDenom := osmoutils.MustExtractDenomFromPacketOnRecv(suite.makeMockPacket("", "", 0))
-	balance := suite.chainA.GetOsmosisApp().BankKeeper.GetBalance(suite.chainA.GetContext(), addr, localDenom)
-	suite.Require().Equal(sdk.NewInt(0), balance.Amount)
-
-	// Execute the contract via IBC with a message that the contract will reject
-	ackBytes := suite.receivePacket(addr.String(), fmt.Sprintf(`{"wasm": {"contract": "%s", "msg": {"not_echo": {"msg": "test"} } } }`, addr))
-	ackStr := string(ackBytes)
-	fmt.Println(ackStr)
-	var ack map[string]string // This can't be unmarshalled to Acknowledgement because it's fetched from the events
-	err := json.Unmarshal(ackBytes, &ack)
-	suite.Require().NoError(err)
-	suite.Require().Contains(ack, "error")
-
-	// Check that the token has now been transferred to the contract
-	balance = suite.chainA.GetOsmosisApp().BankKeeper.GetBalance(suite.chainA.GetContext(), addr, localDenom)
-	fmt.Println(balance)
-	suite.Require().Equal(sdk.NewInt(0), balance.Amount)
-}
-
-func (suite *HooksTestSuite) TestPacketsThatShouldBeSkipped() {
-	var sequence uint64
-	receiver := suite.chainB.SenderAccount.GetAddress().String()
-
-	testCases := []struct {
-		memo           string
-		expPassthrough bool
-	}{
-		{"", true},
-		{"{01]", true}, // bad json
-		{"{}", true},
-		{`{"something": ""}`, true},
-		{`{"wasm": "test"}`, false},
-		{`{"wasm": []`, true}, // invalid top level JSON
-		{`{"wasm": {}`, true}, // invalid top level JSON
-		{`{"wasm": []}`, false},
-		{`{"wasm": {}}`, false},
-		{`{"wasm": {"contract": "something"}}`, false},
-		{`{"wasm": {"contract": "osmo1clpqr4nrk4khgkxj78fcwwh6dl3uw4epasmvnj"}}`, false},
-		{`{"wasm": {"msg": "something"}}`, false},
-		// invalid receiver
-		{`{"wasm": {"contract": "osmo1clpqr4nrk4khgkxj78fcwwh6dl3uw4epasmvnj", "msg": {}}}`, false},
-		// msg not an object
-		{fmt.Sprintf(`{"wasm": {"contract": "%s", "msg": 1}}`, receiver), false},
-	}
-
-	for _, tc := range testCases {
-		ackBytes := suite.receivePacketWithSequence(receiver, tc.memo, sequence)
-		ackStr := string(ackBytes)
-		fmt.Println(ackStr)
-		var ack map[string]string // This can't be unmarshalled to Acknowledgement because it's fetched from the events
-		err := json.Unmarshal(ackBytes, &ack)
-		suite.Require().NoError(err)
-		if tc.expPassthrough {
-			suite.Require().Equal("AQ==", ack["result"], tc.memo)
-		} else {
-			suite.Require().Contains(ackStr, "error", tc.memo)
-		}
-		sequence += 1
-	}
-}
-
-// After successfully executing a wasm call, the contract should have the funds sent via IBC
-func (suite *HooksTestSuite) TestFundTracking() {
-	// Setup contract
-	suite.chainA.StoreContractCode(&suite.Suite, "./bytecode/counter.wasm")
-	addr := suite.chainA.InstantiateContract(&suite.Suite, `{"count": 0}`, 1)
-
-	// Check that the contract has no funds
-	localDenom := osmoutils.MustExtractDenomFromPacketOnRecv(suite.makeMockPacket("", "", 0))
-	balance := suite.chainA.GetOsmosisApp().BankKeeper.GetBalance(suite.chainA.GetContext(), addr, localDenom)
-	suite.Require().Equal(sdk.NewInt(0), balance.Amount)
-
-	// Execute the contract via IBC
-	suite.receivePacket(
-		addr.String(),
-		fmt.Sprintf(`{"wasm": {"contract": "%s", "msg": {"increment": {} } } }`, addr))
-
-	state := suite.chainA.QueryContract(
-		&suite.Suite, addr,
-		[]byte(fmt.Sprintf(`{"get_count": {"addr": "%s"}}`, ibchooks.WasmHookModuleAccountAddr)))
-	suite.Require().Equal(`{"count":0}`, state)
-
-	state = suite.chainA.QueryContract(
-		&suite.Suite, addr,
-		[]byte(fmt.Sprintf(`{"get_total_funds": {"addr": "%s"}}`, ibchooks.WasmHookModuleAccountAddr)))
-	suite.Require().Equal(`{"total_funds":[{"denom":"ibc/C053D637CCA2A2BA030E2C5EE1B28A16F71CCB0E45E8BE52766DC1B241B77878","amount":"1"}]}`, state)
-
-	suite.receivePacketWithSequence(
-		addr.String(),
-		fmt.Sprintf(`{"wasm": {"contract": "%s", "msg": {"increment": {} } } }`, addr), 1)
-
-	state = suite.chainA.QueryContract(
-		&suite.Suite, addr,
-		[]byte(fmt.Sprintf(`{"get_count": {"addr": "%s"}}`, ibchooks.WasmHookModuleAccountAddr)))
-	suite.Require().Equal(`{"count":1}`, state)
-
-	state = suite.chainA.QueryContract(
-		&suite.Suite, addr,
-		[]byte(fmt.Sprintf(`{"get_total_funds": {"addr": "%s"}}`, ibchooks.WasmHookModuleAccountAddr)))
-	suite.Require().Equal(`{"total_funds":[{"denom":"ibc/C053D637CCA2A2BA030E2C5EE1B28A16F71CCB0E45E8BE52766DC1B241B77878","amount":"2"}]}`, state)
-
-	// Check that the token has now been transferred to the contract
-	balance = suite.chainA.GetOsmosisApp().BankKeeper.GetBalance(suite.chainA.GetContext(), addr, localDenom)
-	suite.Require().Equal(sdk.NewInt(2), balance.Amount)
-}
-
-// custom MsgTransfer constructor that supports Memo
-func NewMsgTransfer(
-	token sdk.Coin, sender, receiver string, memo string,
-) *transfertypes.MsgTransfer {
-	return &transfertypes.MsgTransfer{
-		SourcePort:       "transfer",
-		SourceChannel:    "channel-0",
-		Token:            token,
-		Sender:           sender,
-		Receiver:         receiver,
-		TimeoutHeight:    clienttypes.NewHeight(0, 100),
-		TimeoutTimestamp: 0,
-		Memo:             memo,
-	}
-}
-
-type Direction int64
-
-const (
-	AtoB Direction = iota
-	BtoA
-)
-
-func (suite *HooksTestSuite) GetEndpoints(direction Direction) (sender *ibctesting.Endpoint, receiver *ibctesting.Endpoint) {
-	switch direction {
-	case AtoB:
-		sender = suite.path.EndpointA
-		receiver = suite.path.EndpointB
-	case BtoA:
-		sender = suite.path.EndpointB
-		receiver = suite.path.EndpointA
-	}
-	return sender, receiver
-}
-
-func (suite *HooksTestSuite) RelayPacket(packet channeltypes.Packet, direction Direction) (*sdk.Result, []byte) {
-	sender, receiver := suite.GetEndpoints(direction)
-
-	err := receiver.UpdateClient()
-	suite.Require().NoError(err)
-
-	// receiver Receives
-	receiveResult, err := receiver.RecvPacketWithResult(packet)
-	suite.Require().NoError(err)
-
-	ack, err := ibctesting.ParseAckFromEvents(receiveResult.GetEvents())
-	suite.Require().NoError(err)
-
-	// sender Acknowledges
-	err = sender.AcknowledgePacket(packet, ack)
-	suite.Require().NoError(err)
-
-	err = sender.UpdateClient()
-	suite.Require().NoError(err)
-	err = receiver.UpdateClient()
-	suite.Require().NoError(err)
-
-	return receiveResult, ack
-}
-
-func (suite *HooksTestSuite) FullSend(msg sdk.Msg, direction Direction) (*sdk.Result, *sdk.Result, string, error) {
-	var sender *osmosisibctesting.TestChain
-	switch direction {
-	case AtoB:
-		sender = suite.chainA
-	case BtoA:
-		sender = suite.chainB
-	}
-	sendResult, err := sender.SendMsgsNoCheck(msg)
-	suite.Require().NoError(err)
-
-	packet, err := ibctesting.ParsePacketFromEvents(sendResult.GetEvents())
-	suite.Require().NoError(err)
-
-	receiveResult, ack := suite.RelayPacket(packet, direction)
-
-	return sendResult, receiveResult, string(ack), err
-}
-
-func (suite *HooksTestSuite) TestAcks() {
-	suite.chainA.StoreContractCode(&suite.Suite, "./bytecode/counter.wasm")
-	addr := suite.chainA.InstantiateContract(&suite.Suite, `{"count": 0}`, 1)
-
-	// Generate swap instructions for the contract
-	callbackMemo := fmt.Sprintf(`{"ibc_callback":"%s"}`, addr)
-	// Send IBC transfer with the memo with crosschain-swap instructions
-	transferMsg := NewMsgTransfer(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(1000)), suite.chainA.SenderAccount.GetAddress().String(), addr.String(), callbackMemo)
-	suite.FullSend(transferMsg, AtoB)
-
-	// The test contract will increment the counter for itself every time it receives an ack
-	state := suite.chainA.QueryContract(
-		&suite.Suite, addr,
-		[]byte(fmt.Sprintf(`{"get_count": {"addr": "%s"}}`, addr)))
-	suite.Require().Equal(`{"count":1}`, state)
-
-	suite.FullSend(transferMsg, AtoB)
-	state = suite.chainA.QueryContract(
-		&suite.Suite, addr,
-		[]byte(fmt.Sprintf(`{"get_count": {"addr": "%s"}}`, addr)))
-	suite.Require().Equal(`{"count":2}`, state)
-
-}
-
-func (suite *HooksTestSuite) TestSendWithoutMemo() {
-	// Sending a packet without memo to ensure that the ibc_callback middleware doesn't interfere with a regular send
-	transferMsg := NewMsgTransfer(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(1000)), suite.chainA.SenderAccount.GetAddress().String(), suite.chainA.SenderAccount.GetAddress().String(), "")
-	_, _, ack, err := suite.FullSend(transferMsg, AtoB)
-	suite.Require().NoError(err)
-	suite.Require().Contains(ack, "result")
-}
